separate scripts to build, pack, compress & release/publish
